Add tide standards, datums, and platform type
Adds various features that close gulfofmaine/NERACOOS_WaterLevelViewer#11

Tidal datums work by adding the individual datums to a specific time series. That way we can calculate the offset from it without having to reference it from elsewhere. The datums are exposed via the API on the timeseries, but only the configured ones are shown, and if there are none set on a timeseries there is an empty dict.

- Tidal datums closes gulfofmaine/NERACOOS_WaterLevelViewer#19
- Platform type closes gulfofmaine/NERACOOS_WaterLevelViewer#20
- Tidal standard names Closes gulfofmaine/NERACOOS_WaterLevelViewer#21