EDR for FVCOM and BIO forecasts
Uses the data.neracoos.org STAC catalog and Xpublish OGC EDR API to provide forecast data for FVCOM and BIO forecasts.

This removes BIO from our old ERDDAP server, and makes it so that FVCOM no longer depends on UMassD's THREDDS server. Instead it depends on the same processing and access method we are using for the model viewer. We could move to traversing the STAC catalogs and accessing in the front end, but this allows us to retain the flexibility of access for multiple datasources that aren't in the STAC catalog or have an EDR endpoint.

Additionally it makes sure that returned forecasts are cut off a day previous.

Xref: https://github.com/gulfofmaine/NERACOOS-operations/issues/166