Pulls next week of forecast or prediction data
For forecast or prediction timeseries, the next weeks worth of data is pulled instead of the last days, so we can get the extrema from that window. Additionally we also grab the next value, rather than the last one, so it's closer to the current time.

We probably want to change that to a shorter window though.