Admin updates and periodic refresh
This adds a healthcheck to the periodic refresh, so that we know if it isn't firing as expected (which caused a lot of data to go out of date when it's memory needs to launch increased).

This also includes constraining max time requested when a timeseries isn't a forecast, so we don't try to pull from too far into the future.