Admin timeseries and dataset status
This adds columns to the admin with timeseries and dataset refresh status info. 

The columns are color coded, so green if the data is fresh in the last hour, yellow if fresh in the last day, and red otherwise, with an indiciation of how many timeseries have an end time or are marked as not active.

The status columns also support hovering over on both the dataset and platform views to see the refresh time of the dataset, or of each timeseries (seperate active/inactive hovers for timeseries).

This will hopefully make it easier to debug a few different issues we have seen with datasets.