Finetune per-website visibility
Previously `active` and various Mariners/Climatology/Graphing and Download visibilities were a bit more conflated. This cleans those up some. 

Now there is no longer an `active` setting per platform, and the API endpoint instead will use the platforms visibility to toggle if it should be shown. In the case of graphing and download, it should also show all timeseries, not just the active ones.

The climatology visibility is set to false for new platfors now, and needs to be manually turned on.

The `active` setting has now been migrated to Mariner’s visibility.